added counter for block ID

+++ llvm_mode/afl-llvm-rt.o.c
@@ -57,10 +57,10 @@
    is used for instrumentation output before __afl_map_shm() has a chance to run.
    It will end up as .comm, so it shouldn't be too wasteful. */
 
-u8  __afl_area_initial[MAP_SIZE];
-u8* __afl_area_ptr = __afl_area_initial;
+u64  __afl_area_initial[MAP_SIZE];
+u64* __afl_area_ptr = __afl_area_initial;
 
-__thread u32 __afl_prev_loc;
+__thread u64 __afl_prev_loc;
 
 
 /* Running in persistent mode? */
